Xla compilation

+++ compile/mass_compile.py
@@ -24,22 +24,24 @@ class MetaCompile:
         self.frac_compile = frac_compile
         self.compile_list = compile_list
 
+        self.output = {}
+
         if os.path.exists(build_root_dir):
             shutil.rmtree(build_root_dir)
-        os.makedirs(build_root_dir)
 
         if os.path.exists(err_summary_file):
             os.remove(err_summary_file)
 
         if os.path.exists(err_full_info_dir):
             shutil.rmtree(err_full_info_dir)
-        os.makedirs(err_full_info_dir)
 
     def get_compile_list(self):
         if not self.compile_list:
             model_names = [os.path.splitext(file_name)[0]
                            for file_name in os.listdir(self.onnx_model_dir)
-                           if file_name != 'seed.onnx'][::self.frac_compile]
+                           if file_name != 'seed.onnx']
+            model_names.sort(key=lambda x: int(x))
+            model_names = model_names[::self.frac_compile]
             model_names.append('seed')
         else:
             model_names = [str(name) for name in self.compile_list]
@@ -47,57 +49,64 @@ class MetaCompile:
                 model_names.append('seed')
         return model_names
 
-    def compile(self):
-        time_file = os.path.join(self.time_rec_dir, "compile_time.txt")
+    def handle_compilation_error(self, e: CompilationError, model_name):
+        with open(self.err_summary_file, 'a') as f:
+            f.write(f"{e.model_path} $$$ {e.err_code}\n")
+        if not os.path.exists(self.err_full_info_dir):
+            os.makedirs(self.err_full_info_dir)
+        with open(os.path.join(self.err_full_info_dir, "%s.txt" % model_name), 'w') as f:
+            f.write(e.err_info)
 
-        model_names = self.get_compile_list()
 
-        it = time_iterator(model_names, time_file)
+    def compile_run(self, input_file):
+        compile_time_file = os.path.join(self.time_rec_dir, "compile_time.txt")
+        run_time_file = os.path.join(self.time_rec_dir, "run_time.txt")
 
-        for model_name in tqdm.tqdm(it):
-            model_path = os.path.join(self.onnx_model_dir, "%s.onnx" % model_name)
-            build_dir = os.path.join(self.build_root_dir, model_name)
-            os.makedirs(build_dir, exist_ok=True)
-            try:
-                it.cal_time(lambda: self.runner.compile(model_path, build_dir))
-            except CompilationError as e:
-                shutil.rmtree(build_dir)
-                with open(self.err_summary_file, 'a') as f:
-                    f.write(f"{e.model_path} $$$ {e.err_code}\n")
-                with open(os.path.join(self.err_full_info_dir, "%s.txt" % model_name), 'w') as f:
-                    f.write(e.err_info)
-
-    def run(self, input_file):
-        time_file = os.path.join(self.time_rec_dir, "run_time.txt")
-
-        it = time_iterator(os.listdir(self.build_root_dir), time_file)
+        build_dir = self.build_root_dir
 
         self.runner.set_input(input_file)
 
-        for dir_name in it:
-            run_dir = os.path.join(self.build_root_dir, dir_name)
+        model_names = self.get_compile_list()
+
+        it = time_iterator(model_names, [compile_time_file, run_time_file])
+
+        for model_name in tqdm.tqdm(it):
+            model_path = os.path.join(self.onnx_model_dir, "%s.onnx" % model_name)
+            if os.path.exists(build_dir):
+                shutil.rmtree(build_dir)
+            os.mkdir(build_dir)
             try:
-                run_time = self.runner.run(run_dir)
-            except RuntimeError as e:
-                shutil.rmtree(run_dir)
-                with open(self.err_summary_file, 'a') as f:
-                    f.write(f"{run_dir} $$$ {str(e)}\n")
+                it.cal_time(0, lambda: self.runner.compile(model_path, build_dir))
+            except CompilationError as e:
+                self.handle_compilation_error(e, model_name)
+                shutil.rmtree(build_dir)
                 continue
-            it.set_time(run_time)
+
+            try:
+                run_time = self.runner.run(build_dir)
+            except RuntimeError as e:
+                with open(self.err_summary_file, 'a') as f:
+                    f.write(f"{build_dir} $$$ {str(e)}\n")
+                shutil.rmtree(build_dir)
+                continue
+            it.set_time(1, run_time)
+
+            print(self.get_output(build_dir))
+            self.output.update({model_name: self.get_output(build_dir)})
+            shutil.rmtree(build_dir)
 
     def compare_output(self):
-        name_list = [dir_name for dir_name in os.listdir(self.build_root_dir)
-                     if dir_name != 'seed']
+        name_list = [model_name for model_name in self.output.keys() if model_name != 'seed']
         name_list.sort(key=lambda x: int(x))
 
-        seed_output = self.get_output('seed')
+        seed_output = self.output['seed']
 
-        diff_list = [array_diff(self.get_output(name), seed_output) for name in name_list]
+        diff_list = [array_diff(self.output[name], seed_output) for name in name_list]
 
         write_output_diff(self.diff_file_path, diff_list, name_list)
 
-    def get_output(self, model_name):
-        return self.runner.get_output(os.path.join(self.build_root_dir, model_name))
+    def get_output(self, build_dir):
+        return self.runner.get_output(build_dir)
 
 
 def compiler_run(compiler_name, compiler_path, onnx_model_dir,
@@ -113,6 +122,5 @@ def compiler_run(compiler_name, compiler_path, onnx_model_dir,
                                 diff_file_path, time_record_dir,
                                 err_summary_file, err_full_info_dir,
                                 frac_compile, compile_list)
-    meta_compiler.compile()
-    meta_compiler.run(data_path)
+    meta_compiler.compile_run(data_path)
     meta_compiler.compare_output()
